@@ -9,9 +9,9 @@
     "axios": "^1.7.0",
     "bootstrap": "^4.6.0",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^1.6.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-infinite-scroll-component": "^6.1.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "^5.0.1",
