@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { test, render, screen, expect } from '@testing-library/react';
 import App from './App';
 
 test('renders learn react link', () => {
