@@ -20,7 +20,9 @@ const NavBar = () => {
       await axios.post("dj-rest-auth/logout/");
       setCurrentUser(null);
       removeTokenTimestamp();
-    } catch (err) {}
+    } catch (err) {
+      // console.log(err);
+    }
   };
 
 
