@@ -3,6 +3,7 @@ import axios from "axios";
 import { axiosReq, axiosRes } from "../api/axiosDefaults";
 import { useHistory } from "react-router";
 import { removeTokenTimestamp, shouldRefreshToken } from "../utils/utils";
+import PropTypes from 'prop-types';
 
 export const CurrentUserContext = createContext();
 export const SetCurrentUserContext = createContext();
@@ -14,12 +15,16 @@ export const CurrentUserProvider = ({ children }) => {
   const [currentUser, setCurrentUser] = useState(null);
   const history = useHistory();
 
+  CurrentUserProvider.propTypes = {
+    children: PropTypes.node.isRequired,
+  };
+
   const handleMount = async () => {
     try {
       const { data } = await axiosRes.get("dj-rest-auth/user/");
       setCurrentUser(data);
     } catch (err) {
-      
+      // console.log(err);
     }
   };
 
