@@ -2,6 +2,7 @@ import { createContext, useContext, useEffect, useState } from "react";
 import { axiosReq, axiosRes } from "../api/axiosDefaults";
 import { useCurrentUser } from "../contexts/CurrentUserContext";
 import { followHelper, unfollowHelper } from "../utils/utils";
+import PropTypes from 'prop-types';
 
 const ProfileDataContext = createContext();
 const SetProfileDataContext = createContext();
@@ -11,10 +12,13 @@ export const useSetProfileData = () => useContext(SetProfileDataContext);
 
 export const ProfileDataProvider = ({ children }) => {
     const [profileData, setProfileData] = useState({
-        // we will use the pageProfile later
         pageProfile: { results: [] },
         popularProfiles: { results: [] },
       });
+
+  ProfileDataProvider.propTypes = {
+    children: PropTypes.node.isRequired,
+  };
 
   const currentUser = useCurrentUser();
 
@@ -39,7 +43,7 @@ export const ProfileDataProvider = ({ children }) => {
         },
       }));
     } catch (err) {
-
+      // console.log(err);
     }
   };
 
@@ -61,7 +65,7 @@ export const ProfileDataProvider = ({ children }) => {
         },
       }));
     } catch (err) {
-    
+      // console.log(err);
     }
   };
 
@@ -76,6 +80,7 @@ export const ProfileDataProvider = ({ children }) => {
           popularProfiles: data,
         }));
       } catch (err) {
+        // console.log(err);
       }
     };
 
