@@ -6,6 +6,7 @@ import { Link, useHistory } from "react-router-dom";
 import Avatar from "../../components/Avatar";
 import { axiosRes } from "../../api/axiosDefaults";
 import { MoreDropdown } from "../../components/MoreDropdown";
+import PropTypes from 'prop-types';
 
 const Post = (props) => {
   const {
@@ -25,7 +26,23 @@ const Post = (props) => {
     setPosts,
   } = props;
 
-  
+  Post.propTypes = {
+    id: PropTypes.string.isRequired,
+    owner: PropTypes.string.isRequired,
+    profile_id: PropTypes.string.isRequired,
+    profile_image: PropTypes.string.isRequired,
+    comments_count: PropTypes.number.isRequired,
+    likes_count: PropTypes.number.isRequired,
+    like_id: PropTypes.string,
+    title: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+    image: PropTypes.string,
+    updated_at: PropTypes.string.isRequired,
+    category_name: PropTypes.string.isRequired,
+    postPage: PropTypes.bool.isRequired,
+    setPosts: PropTypes.func.isRequired,
+  };
+
   const currentUser = useCurrentUser();
   const is_owner = currentUser?.username === owner;
   const history = useHistory();
@@ -39,6 +56,7 @@ const Post = (props) => {
       await axiosRes.delete(`/posts/${id}/`);
       history.goBack();
     } catch (err) {
+      // console.log(err);
     }
   };
 
@@ -106,7 +124,7 @@ const Post = (props) => {
           {is_owner ? (
             <OverlayTrigger
               placement="top"
-              overlay={<Tooltip>You can't like your own post!</Tooltip>}
+              overlay={<Tooltip>You can&apos;t like your own post!</Tooltip>}
             >
               <i className="far fa-heart" />
             </OverlayTrigger>
