@@ -57,7 +57,7 @@ function ProfilePage() {
                 setProfilePosts(profilePosts);
                 setHasLoaded(true);
             } catch (err) {
-                
+                //console.error();
             }
         };
         fetchData();
@@ -189,31 +189,22 @@ function ProfilePage() {
             <p className="text-center">
                 {
                 profile?.owner
-            }'s posts</p>
+            }&apos;s posts</p>
             <hr/> {
             profilePosts.results.length ? (
-                <InfiniteScroll children={
-                        profilePosts.results.map((post) => (
-                            <Post key={
-                                    post.id
-                                }
+                 
+                    <InfiniteScroll
+                    dataLength={profilePosts.results.length}
+                    loader={<Asset spinner/>}
+                    hasMore={!!profilePosts.next}
+                    next={() => fetchMoreData(profilePosts, setProfilePosts)}
+                  > 
+                   {profilePosts.results.map((post) => (
+                        <Post key={post.id}
                                 {...post}
                                 setPosts={setProfilePosts}/>
-                        ))
-                    }
-                    dataLength={
-                        profilePosts.results.length
-                    }
-                    loader={
-                        <Asset
-                    spinner/>
-                    }
-                    hasMore={
-                        !!profilePosts.next
-                    }
-                    next={
-                        () => fetchMoreData(profilePosts, setProfilePosts)
-                    }/>
+                     ))}
+              </InfiniteScroll>
             ) : (
                 <Asset src={NoResults}
                     message={
