@@ -13,24 +13,23 @@ export const fetchMoreData = async (resource, setResource) => {
           : [...acc, cur];
       }, prevResource.results),
     }));
-  } catch (err) {}
+  } catch (err) {
+    //console.error();
+  }
 };
 
 export const followHelper = (profile, clickedProfile, following_id) => {
   return profile.id === clickedProfile.id
-    ? // This is the profile I clicked on,
-      // update its followers count and set its following id
+    ?
       {
         ...profile,
         followers_count: profile.followers_count + 1,
         following_id,
       }
     : profile.is_owner
-    ? // This is the profile of the logged in user
-      // update its following count
+    ? 
       { ...profile, following_count: profile.following_count + 1 }
-    : // this is not the profile the user clicked on or the profile
-      // the user owns, so just return it unchanged
+    :
       profile;
 };
 
